requirements.txt
django==4.2.10
django-cors-headers==4.3.1
djangorestframework==3.14.0
psycopg==3.1.19
dj-database-url==2.1.0
python-dotenv==1.0.0
graphene==3.3
graphene-django==3.2.0